+++ cpp_impl/common.h
@@ -13,6 +13,6 @@
 #include "data_structures/InterleavedTCSCPadding.h"
 #include "data_structures/InterleavedTCSC_baraq.h"
 
-using comp_func = std::function<void(double *X, double *B, double *Y, int M, int N, int K)>;
+using comp_func = std::function<void(float *X, float *B, float *Y, int M, int N, int K)>;
 
 void add_function(comp_func f, std::string name);
